Avoid deprecation warnings. CSI is GA as of 1.13
According to https://kubernetes.io/blog/2019/01/15/container-storage-interface-ga/, CSI became GA in 1.13. As of 1.20, it's throwing deprecation warnings, triggering issue #396 . I'm not sure how all of the build process works for this project, but this appears to be a reasonable place to reflect where GA actually happened so that the latest versions of the driver can stop throwing deprecation warnings.