Bump the livenessProbe sidcare to v2.6.0-eks-1-18-17
Solves #564. There is a proposed solution for this with #595, but
it changes the container registry and is therefore a bigger change
then just go with the next release version (which fixes the issue)
from the same registry as already used.

It might be wise, to just bump all the requirements from currently
1-18-13 to 1-18-17. I am not doing this in this PR, since this
only aims at resolving #564 which is really annoying for us.

For the time of this not being fixed, one can easily replace the
default livenessProbe sidecare version with the following:
https://github.com/kubernetes-sigs/aws-efs-csi-driver/issues/564#issuecomment-1118496793