Splitting up provisioner into the requisite parts
Now we have an AccessPoint and Directory provisioner they should live in different files with their own tests.
We also need an OsClient to stop us having to run the tests as root which could cause issues in CI/CD.

Sets the directory permissions to 777 by default and doesn't set an
owner to fix the problem of how users would know what UID/GID to use.
Further fixes a problem where if an unmount failed it would delete the
contents of the EFS.