allow specifying timeout duration
On an EKS cluster we experienced some node pods entering crashloop while
volumes were actually correctly attached to nodes and mounted on pods.
This was due to the liveness probe taking too long (its default timeout
is 1s), failing with a timeout and not managing to recover. The probe
continued to fail even after the container is restarted.

AWS support suggested increasing timeouts which was performed by manual
updates on the daemonset and pods are running fine since then. But those
timeouts could not be defined in the Helm chart.

Add timeout values in the Helm chart with default values set to their
current default.