Merge the controller and node service accounts
To enforce filesystem policies, efs-utils must be passed the iam option.
However, this also implies that the node-serviceaccount must be
annotated with the IAM Role that has the corresponding filesystem policy
permissions.

When installing the CSI Driver, customers are instructed to annotate the
controller-serviceaccount. Instead of having them additionally annotate
the node-serviceaccount so that their filesystem policies can be adhered
to, we should merge these two service accounts together so that they
only need to annotate one Service Account instead of two.