set results count for listing access points
DescribeAccessPointsWithContext call returns only 100 results by
default.

Since this is used for discovering used GIDs on access points we must
not allow listing only a subset of access points. This is because
gid allocator relies on this call to filter out used GIDs and getting
next *free* GID. If we don't make sure to list every access point it
will cause allocator to pick GID that might already be taken.