Reduce calls to EFS API
With #850, the new way of allocating GIDs introduced a new call
to the ListAccessPoints endpoint of EFS API, that is generating
problems on systems where EFS CSI driver is under high load
(lots of PVC created within a short time period). In this PR,
we are extracting the ListAccessPoints call from gid_allocator,
moving it one level up. In case of dynamic provisioining of GIDs
we can reuse the ListAccessPoints call to check for the file
system existence (thus removing the DescribeFileSystem call in
such case). In case of a fixed UID/GID, we continue calling
DescribeFileSystem, and no calls to ListAccessPoints.

In addition to the change explained above, gidMin and gidMax have
been converted to int64. #850 made both uid and gid int64, but
gidMin and gidMax were nout touched. Also changing the default
value for gidMax, as the value of 7000000 was spamming the logs
with a message coming from gid_allocator (i.e. range bigger than
the max number of access points). Setting the value to 51000
(the value that gid_allocator was setting by adding 1000 to
gidMin). Removing an unused function and field from gid_allocator
too.

(cherry picked from commit dbbc733333eb02195c5e3653546b63778f291cf4)