Refactor re-use Access Point
When reuse access point flag is set to true, there is no point
in doing a lot of work that is aimed at preparing the options for
the createAccessPoint call (e.g. allocating a GID). We can save
quite some time and quite some API calls if the AP with the same
clientToken is found. The check for the AP existence has been moved
from the CreateAccessPoint function one level up. The controller
in this case checks if there is an access point with such client
token, if so, it proceeds to return a CreateVolumeResponse. If it
is not found, then it follows the usual flow.