Allow controller SA to describe x-account Secret
For cross account provisioning, the controller pods need to be able to
describe the x-account K8s Secret in the kube-system namespace.

Additionally, I modified the IAM Role used in our cross-account mount
example. Users should use the managed AmazonEFSCSIDriverPolicy, which
I've copied here. Eventually, we will update the documentation to
reference the managed policy instead of this file.